// src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        Poll: {
            akkoma?: {
                /** @description Whether the poll creator indicates votes are processed anonymously (true), votes with voter identity will be published to some parties or the public (false) or did not indicate anything about vote confidentiality (null) */
                anonymous?: boolean | null;
            };
            /** @description Custom emoji to be used for rendering poll options. */
            emojis?: components["schemas"]["Emoji"][];
            /** @description Is the poll currently expired? */
            expired?: boolean;
            /**
             * Format: date-time
             * @description When the poll ends
             */
            expires_at?: string | null;
            id?: components["schemas"]["FlakeID"];
            /** @description Does the poll allow multiple-choice answers? */
            multiple?: boolean;
            /** @description Possible answers for the poll. */
            options?: {
                title?: string;
                votes_count?: number;
            }[];
            /** @description When called with a user token, has the authorized user voted? Boolean, or null if no current user. */
            voted?: boolean | null;
            /** @description How many unique accounts have voted for a multi-selection poll. Number, or null if single-selection poll. */
            voters_count?: number | null;
            /** @description How many votes have been received. Number. */
            votes_count?: number;
        };
// ...
        Attachment: {
            /** @description Alternate text that describes what is in the media attachment, to be used for the visually impaired or when media attachments do not load */
            description?: string | null;
            /** @description The ID of the attachment in the database. */
            id?: string;
            pleroma?: {
                /** @description mime type of the attachment */
                mime_type?: string;
            };
            /**
             * Format: uri
             * @description The location of a scaled-down preview of the attachment
             */
            preview_url?: string;
            /**
             * Format: uri
             * @description The location of the full-size original attachment on the remote website. String (URL), or null if the attachment is local
             */
            remote_url?: string | null;
            /**
             * Format: uri
             * @description A shorter URL for the attachment
             */
            text_url?: string;
            /**
             * @description The type of the attachment
             * @enum {string}
             */
            type?: "image" | "video" | "audio" | "unknown";
            /**
             * Format: uri
             * @description The location of the original full-size attachment
             */
            url?: string;
        };
// ...
    };
    responses: never;
    parameters: {
        /**
         * @description Account ID or nickname
         * @example 123
         */
        accountIdOrNickname: string;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}